feed write tab 제거 및 하단 floatbutton 클릭시 이동
