+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Thirdctivity" />
+        <activity android:name=".FeedWriteActivity" />
         <activity android:name=".ui.set.SecondActivity" />
         <activity
             android:name=".ChatActivity"
