+++ app/src/main/res/layout/activity_feedwrite.xml
