+++ app/src/main/res/layout/app_bar_main.xml
@@ -28,6 +28,6 @@
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
         android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@android:drawable/ic_dialog_email" />
+        app:srcCompat="@drawable/ic_baseline_control_point_24" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
