--- app/src/main/res/layout/fragment_feed.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.home.FeedListFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/feed_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true">
-
-    </androidx.recyclerview.widget.RecyclerView>
-
-</RelativeLayout>
