--- app/src/main/res/layout/fragment_feedwrite.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_marginTop="30dp"
-        android:layout_marginBottom="30dp">
-
-        <ImageView
-            android:id="@+id/feed_veiw"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:scaleType="fitCenter"
-            android:layout_gravity="center_vertical|center_horizontal"
-            app:srcCompat="@drawable/no_change" />
-
-        <Button
-            android:id="@+id/btn_change_photo"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_gravity="center"
-            android:text="+"></Button>
-
-    </LinearLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="2dp"
-        android:layout_centerInParent="true"
-        android:background="#dfdfdf"/>
-
-
-
-    <EditText
-        android:layout_width="match_parent"
-        android:layout_height="70dp"
-        android:hint="Text.."
-        android:background="@color/white">
-    </EditText>
-
-
-
-</LinearLayout>
