+++ app/src/main/res/layout/fragment_home.xml
@@ -1,22 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/main_tabs"
-        android:background="@drawable/white_gray_border_bottom"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".ui.home.HomeFragment">
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/feed_list"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="@style/AppTheme"/>
+        android:layout_height="match_parent"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true">
 
-    <androidx.viewpager.widget.ViewPager
-        android:id="@+id/tabs_pager"
-        android:layout_below="@id/main_tabs"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+    </androidx.recyclerview.widget.RecyclerView>
 
-    </androidx.viewpager.widget.ViewPager>
-</LinearLayout>
+</RelativeLayout>
